Fix intermittent issues with keyboard rebinds
Sometimes they would error on initial load which I thought
was during the shortcuts being removed from spicetify 2.20
since they had been added to spotify by that point but just
replacing the depricated keyboard library in favour of the
newer mousetrap one appears to have solved it.
Since mod+/ is actually used in spotify I have added a check
for the time being to make sure it doesn't replace mod+k on
startup if the setting is not enabled. It might be worth just
changing the shortcut to one that isn't used but at a minumum
the check should prevent any future issues if new shortcuts
are added to either spotify or spicetify.